fix JR instr, romBank

+++ SwiftyBoy/SwiftyBoy/GameBoy/RAM.swift
@@ -9,10 +9,10 @@ import Foundation
 
 class RAM {
     
-    var internalRam0 = Array<UInt8>(repeating: 0x00, count: 8 * 1024)
-    var nonIoInternalRam0 = Array<UInt8>(repeating: 0x00, count: 0x60)
-    var ioPortsRAM = Array<UInt8>(repeating: 0x00, count: 0x4C)
-    var internalRam1 = Array<UInt8>(repeating: 0x00, count: 0x7F)
-    var nonIoInternalRam1 = Array<UInt8>(repeating: 0x00, count: 0x34)
+    var internalRam0 = Array<Int>(repeating: 0x00, count: 8 * 1024)
+    var nonIoInternalRam0 = Array<Int>(repeating: 0x00, count: 0x60)
+    var ioPortsRAM = Array<Int>(repeating: 0x00, count: 0x4C)
+    var internalRam1 = Array<Int>(repeating: 0x00, count: 0x7F)
+    var nonIoInternalRam1 = Array<Int>(repeating: 0x00, count: 0x34)
     
 }
